ml/myappzip/requirements.txt
google-cloud-tasks==1.5.0
googleapis-common-protos==1.51.0
Flask==1.1.1
SQLAlchemy==1.3.16
PyMySQL==0.9.3
pandas
sklearn
